delete GWMcorrelation.scala
